+++ routes/elections/updateElection.ts
@@ -5,9 +5,9 @@ import HttpError from '../../httpError';
 import { decodeEntityAndVerifyAccess } from '../../utils';
 
 const updateElectionSchema= z.object({
-  group_id: z.number(),
-  user_id: z.number(),
-  election_id: z.number(),
+  group_id: z.number().int(),
+  user_id: z.number().int(),
+  election_id: z.number().int(),
   status: z.enum(['Open', 'Cancelled']).optional().nullable(),
   nomination_ends_at: z.string().datetime().optional().nullable()
 });
@@ -26,8 +26,8 @@ const updateElections = (router: Router) => {
     description: 'Allows updating election status and nomination end date if not closed',
     request: {
       params: z.object({
-        group_id: z.number(),
-        election_id: z.number()
+        group_id: z.number().int(),
+        election_id: z.number().int()
       }),
       body: z.object({
         status: z.enum(['Open', 'Cancelled']).optional(),
